+++ packages/support/src/mixins/decorators/Bare.ts
