+++ packages/support/src/mixins/index.ts
@@ -1,4 +1,4 @@
-export * from './mixins';
+export * from './decorators';
 export * from './apply';
 export * from './hasMixin';
 export * from './isApplicationOf';
